+++ src/modules/next-graded-course.js
@@ -6,6 +6,10 @@ import {
 } from '~/utils/dom';
 import { coursesListLoaded } from '~/shared/progress';
 
+const selectors = {
+  button: 'gocp_next-graded-course_button',
+};
+
 function selectCourse({ elem }) {
   elem.parentNode.querySelector('.btn.btn-default').click();
 }
@@ -14,6 +18,7 @@ function generateButton(followingCourse, text) {
   const button = constructButton(
     text, '', '',
     () => selectCourse(followingCourse),
+    selectors.button,
   );
   button.style.color = '';
 
@@ -35,6 +40,23 @@ function getPreviousCourse(courses, course) {
   return getNextCourse(reversedCourses, course);
 }
 
+async function addNextGradedCourseButtons(courses, currentCourse, unloaderContext) {
+  await waitForLoad(() => document.querySelectorAll('button[data-analysis="next"]').length);
+
+  const nextCourse = getNextCourse(courses, currentCourse);
+  const prevCourse = getPreviousCourse(courses, currentCourse);
+  const nextGradedButton = generateButton(nextCourse, 'Next Graded Course');
+  const prevGradedButton = generateButton(prevCourse, 'Previous Graded Course');
+
+  const nextButton = document.querySelectorAll('button[data-analysis="next"]')[0];
+  const prevButton = document.querySelectorAll('button[data-analysis="prev"]')[0];
+  nextButton.after(nextGradedButton);
+  prevButton.before(prevGradedButton);
+
+  unloaderContext.addRemovable(nextGradedButton);
+  unloaderContext.addRemovable(prevGradedButton);
+}
+
 const domQuery = () => (
   coursesListLoaded() &&
   document.querySelector('.btn.btn-default.btn-sm.bold')
@@ -52,22 +74,25 @@ async function nextGradedCourse(opts, unloaderContext) {
   const courses = Array.from(document.getElementsByClassName('showGrade'))
     .map(gradeElemToObject);
 
+  const reAddButtons = (currentCourse, direction) => {
+    const buttons = document.querySelectorAll(`.${selectors.button}`);
+    for (const button of buttons) {
+      button.remove();
+    }
+    const followingCourse = courses[courses.indexOf(currentCourse) + direction];
+    addNextGradedCourseButtons(courses, followingCourse, unloaderContext);
+  };
+
   for (const course of courses) {
     const gradeDetailButton = course.elem.parentNode.querySelector('.btn.btn-default');
     const gradeDetailListener = addEventListener(gradeDetailButton, 'click', async () => {
-      await waitForLoad(() => document.querySelectorAll('button[data-analysis="next"]').length);
-      const nextCourses = getNextCourse(courses, course);
-      const prevCourses = getPreviousCourse(courses, course);
-      const nextGradedButton = generateButton(nextCourses, 'forward');
-      const prevGradedButton = generateButton(prevCourses, 'backward');
+      await addNextGradedCourseButtons(courses, course, unloaderContext);
 
       const nextButton = document.querySelectorAll('button[data-analysis="next"]')[0];
       const prevButton = document.querySelectorAll('button[data-analysis="prev"]')[0];
-      nextButton.after(nextGradedButton);
-      prevButton.before(prevGradedButton);
 
-      unloaderContext.addRemovable(nextGradedButton);
-      unloaderContext.addRemovable(prevGradedButton);
+      unloaderContext.addRemovable(addEventListener(nextButton, 'click', () => reAddButtons(course, 1)));
+      unloaderContext.addRemovable(addEventListener(prevButton, 'click', () => reAddButtons(course, -1)));
     });
 
     unloaderContext.addRemovable(gradeDetailListener);